projects/demand-forecasting/requirements.txt
# Walmart Demand Forecasting - Requirements
# Autor: Ing. Daniel Varela Perez
# Email: bedaniele0@gmail.com
# Python: 3.10+

# ============================================
# CORE DATA PROCESSING
# ============================================
pandas==2.2.0
numpy==1.26.0
scipy==1.11.0

# ============================================
# TIME SERIES & FORECASTING
# ============================================
statsmodels==0.14.0
prophet==1.1.0
pmdarima==2.0.0
sktime==0.25.0

# ============================================
# MACHINE LEARNING
# ============================================
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.0

# ============================================
# DEEP LEARNING (Optional - for LSTM/GRU)
# ============================================
# tensorflow==2.15.0
# torch==2.1.2

# ============================================
# OPTIMIZATION
# ============================================
pulp==2.7.0

# ============================================
# VISUALIZATION
# ============================================
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0

# ============================================
# MLOPS & EXPERIMENT TRACKING
# ============================================
mlflow==2.9.0

# ============================================
# API & DASHBOARD
# ============================================
fastapi==0.108.0
uvicorn[standard]==0.25.0
streamlit==1.29.0
pydantic==2.5.0

# ============================================
# DATA VALIDATION & QUALITY
# ============================================
pandera==0.18.0
great-expectations==0.18.0

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
pyyaml==6.0.0
joblib==1.3.0
tqdm==4.66.0
click==8.1.0

# ============================================
# TESTING
# ============================================
pytest==7.4.0
pytest-cov==4.1.0

# ============================================
# CODE QUALITY
# ============================================
black==23.12.0
flake8==7.0.0
isort==5.13.0

# ============================================
# JUPYTER & NOTEBOOKS
# ============================================
jupyter==1.0.0
ipykernel==6.27.0
ipywidgets==8.1.0
notebook==7.0.0

# ============================================
# DOCUMENTATION
# ============================================
sphinx==7.2.0
sphinx-rtd-theme==2.0.0

# ============================================
# MONITORING & LOGGING
# ============================================
loguru==0.7.0

# ============================================
# EXTRAS
# ============================================
requests==2.31.0
python-dateutil==2.8.0
pytz==2023.3
